jit: Break out architecture-agnostic parts to prepare for ARM JIT
