+++ erts/emulator/beam/jit/beam_jit_common.hpp
@@ -18,6 +18,9 @@
  * %CopyrightEnd%
  */
 
+#ifndef __BEAM_JIT_COMMON_HPP__
+#define __BEAM_JIT_COMMON_HPP__
+
 #include <string>
 #include <vector>
 #include <unordered_map>
@@ -30,10 +33,10 @@ extern "C"
 #endif
 
 #include "sys.h"
+#include "bif.h"
 #include "erl_vm.h"
 #include "global.h"
-
-#include "beam_jit_common.h"
+#include "beam_file.h"
 }
 
 class ArgVal {
@@ -135,3 +138,83 @@ public:
         }
     };
 };
+
+/* ** */
+
+#if defined(DEBUG) && defined(JIT_HARD_DEBUG)
+void beam_jit_validate_term(Eterm term);
+#endif
+
+typedef Eterm BeamJitNifF(struct enif_environment_t *, int argc, Eterm argv[]);
+enum beam_jit_nif_load_ret { RET_NIF_success, RET_NIF_error, RET_NIF_yield };
+
+Eterm beam_jit_call_bif(Process *c_p,
+                        Eterm *reg,
+                        ErtsCodePtr I,
+                        ErtsBifFunc vbf,
+                        Uint arity);
+Eterm beam_jit_call_nif(Process *c_p,
+                        ErtsCodePtr I,
+                        Eterm *reg,
+                        BeamJitNifF *fp,
+                        struct erl_module_nif *NifMod);
+enum beam_jit_nif_load_ret beam_jit_load_nif(Process *c_p,
+                                             ErtsCodePtr I,
+                                             Eterm *reg);
+
+Uint beam_jit_get_map_elements(Eterm map,
+                               Eterm *reg,
+                               Eterm *E,
+                               Uint n,
+                               Eterm *fs);
+
+void beam_jit_bs_field_size_argument_error(Process *c_p, Eterm size);
+void beam_jit_bs_add_argument_error(Process *c_p, Eterm A, Eterm B);
+Eterm beam_jit_bs_start_match2(Eterm context,
+                               Uint live,
+                               Uint slots,
+                               Process *c_p,
+                               Eterm *reg);
+Eterm beam_jit_bs_init(Process *c_p,
+                       Eterm *reg,
+                       ERL_BITS_DECLARE_STATEP,
+                       Eterm BsOp1,
+                       Eterm BsOp2,
+                       unsigned Live);
+Eterm beam_jit_bs_init_bits(Process *c_p,
+                            Eterm *reg,
+                            ERL_BITS_DECLARE_STATEP,
+                            Uint num_bits,
+                            Uint alloc,
+                            unsigned Live);
+Eterm beam_jit_bs_get_integer(Process *c_p,
+                              Eterm *reg,
+                              Eterm context,
+                              Uint flags,
+                              Uint size,
+                              Uint Live);
+
+ErtsMessage *beam_jit_decode_dist(Process *c_p, ErtsMessage *msgp);
+Sint beam_jit_remove_message(Process *c_p,
+                             Sint FCALLS,
+                             Eterm *HTOP,
+                             Eterm *E,
+                             Uint32 active_code_ix);
+void beam_jit_bs_context_to_binary(Eterm context);
+
+void beam_jit_take_receive_lock(Process *c_p);
+void beam_jit_wait_locked(Process *c_p, ErtsCodePtr cp);
+void beam_jit_wait_unlocked(Process *c_p, ErtsCodePtr cp);
+
+enum beam_jit_tmo_ret { RET_next = 0, RET_wait = 1, RET_badarg = 2 };
+
+enum beam_jit_tmo_ret beam_jit_wait_timeout(Process *c_p,
+                                            Eterm timeout_value,
+                                            ErtsCodePtr next);
+
+void beam_jit_timeout(Process *c_p);
+void beam_jit_timeout_locked(Process *c_p);
+
+void beam_jit_return_to_trace(Process *c_p);
+
+#endif
